Migrate to GitHub Actions
Heavily copied code from both
https://github.com/dedupeio/hcluster/blob/59bc901288bac93a5f61bbe5bab2033b117d00e1/.github/workflows/pythonpackage.yml
and
https://github.com/pypa/cibuildwheel/blob/main/examples/github-deploy.yml

Not totally sure about the configuration of cibuildwheel.
I tried to build it for as many architectures on linux as possible
(using emulation with QEMU) and as many architectures as possible
on MacOS. Only bothered with tje native architecture on Windows.
Not sure if I did that config correctly.

I *think* `auditwheel repair` is getting run in the `cibuildwheel`
action, but not sure?

I disabled building any wheels for pypy (the `pp*` in CIBW_SKIP).
Maybe they could get added back in.

This adds testing for python 3.10, which wasn't there before.
This also always builds wheels on every push and PR, it just
only uploads them to PyPI on tagging. This is different from
before, when building only happened on tags.

Also, automatically publish a release to GitHub in addition to
publishing to PyPI.